+++ test/config.spec.js
@@ -2,7 +2,7 @@ const Joi = require('joi');
 
 const nullable = schema => schema.optional().allow(null);
 
-const configScheme = {
+const configScheme = Joi.object({
   context: Joi.string(),
   protocol: Joi.any().valid('sftp', 'ftp', 'test'),
 
@@ -14,11 +14,9 @@ const configScheme = {
   agent: nullable(Joi.string()),
   privateKeyPath: nullable(Joi.string()),
   passphrase: nullable(Joi.string().allow(true)),
-  interactiveAuth: Joi.alternatives([
-    Joi.boolean(),
-    Joi.array()
-      .items(Joi.string()),
-  ]).optional(),
+  interactiveAuth: Joi.alternatives()
+    .try(Joi.boolean(), Joi.array().items(Joi.string()))
+    .optional(),
 
   secure: Joi.any().valid(true, false, 'control', 'implicit').optional(),
   secureOptions: nullable(Joi.object()),
@@ -39,7 +37,7 @@ const configScheme = {
     autoUpload: Joi.boolean().optional(),
     autoDelete: Joi.boolean().optional(),
   },
-};
+});
 
 describe("validation config", () => {
   test("default config", () => {
@@ -75,10 +73,10 @@ describe("validation config", () => {
       ],
     };
 
-    const result = Joi.validate(config, configScheme, {
+    const result = configScheme.validate(config, {
       convert: false,
     });
-    expect(result.error).toBe(null);
+    expect(result.error).toBeUndefined();
   });
 
   test("partial config", () => {
@@ -101,16 +99,16 @@ describe("validation config", () => {
       ],
     };
 
-    let result = Joi.validate(config, configScheme, {
+    let result = configScheme.validate(config, {
       convert: false,
     });
-    expect(result.error).toBe(null);
+    expect(result.error).toBeUndefined();
 
     delete config.watcher;
-    result = Joi.validate(config, configScheme, {
+    result = configScheme.validate(config, {
       convert: false,
     });
-    expect(result.error).toBe(null);
+    expect(result.error).toBeUndefined();
   });
 
   describe("key validaiton", () => {
@@ -144,10 +142,10 @@ describe("validation config", () => {
         ],
       };
 
-      const result = Joi.validate(config, configScheme, {
+      const result = configScheme.validate(config, {
         convert: false,
       });
-      expect(result.error).not.toBe(null);
+      expect(result.error).toBeDefined();
     });
 
     test("watcher files must be false or string", () => {
@@ -180,34 +178,34 @@ describe("validation config", () => {
         ],
       };
 
-      let result = Joi.validate(config, configScheme, {
+      let result = configScheme.validate(config, {
         convert: false,
       });
-      expect(result.error).toBe(null);
+      expect(result.error).toBeUndefined();
 
       config.watcher.files = '**/*.js';
-      result = Joi.validate(config, configScheme, {
+      result = configScheme.validate(config, {
         convert: false,
       });
-      expect(result.error).toBe(null);
+      expect(result.error).toBeUndefined();
 
       config.watcher.files = null;
-      result = Joi.validate(config, configScheme, {
+      result = configScheme.validate(config, {
         convert: false,
       });
-      expect(result.error).toBe(null);
+      expect(result.error).toBeUndefined();
 
       config.watcher.files = true;
-      result = Joi.validate(config, configScheme, {
+      result = configScheme.validate(config, {
         convert: false,
       });
-      expect(result.error).not.toBe(null);
+      expect(result.error).toBeDefined();
 
       delete config.watcher;
-      result = Joi.validate(config, configScheme, {
+      result = configScheme.validate(config, {
         convert: false,
       });
-      expect(result.error).toBe(null);
+      expect(result.error).toBeUndefined();
     });
 
     test("ignore must be an array of string", () => {
@@ -240,16 +238,16 @@ describe("validation config", () => {
         ],
       };
 
-      let result = Joi.validate(config, configScheme, {
+      let result = configScheme.validate(config, {
         convert: false,
       });
-      expect(result.error).not.toBe(null);
+      expect(result.error).toBeDefined();
 
       config.ignore = [];
-      result = Joi.validate(config, configScheme, {
+      result = configScheme.validate(config, {
         convert: false,
       });
-      expect(result.error).toBe(null);
+      expect(result.error).toBeUndefined();
     });
 
     test("pass", () => {
@@ -282,16 +280,16 @@ describe("validation config", () => {
         ],
       };
 
-      let result = Joi.validate(config, configScheme, {
+      let result = configScheme.validate(config, {
         convert: false,
       });
-      expect(result.error).toBe(null);
+      expect(result.error).toBeUndefined();
 
       config.passphrase = false;
-      result = Joi.validate(config, configScheme, {
+      result = configScheme.validate(config, {
         convert: false,
       });
-      expect(result.error).not.toBe(null);
+      expect(result.error).toBeDefined();
     });
   });
 });